fix: 마우스 스크롤 시 진행률 반영
